Add a recipe page
We add a folder and inside it an html page with the
name of our recipe.

We fill our recipe page with the usual boilerplate as
well and a heading h1 with the name of our recipe.

We also add a link from our index.html to the recipe
page and a link from the recipe page to our index.